feature(docker): split volumes & add traefik & fix logging
We want to split the volumes. We want blobstorage, filestorage and logs
to be persistent. Only blob and logs will be mounted in the instance to
share the blob between the server and the instance.

Traefik will enable to do balancing between the instances.

We also fix the logging to have separate logging files which are named
based on the container hostname.